Create commands/queries for users and accessory

+++ Ghini-Bikes/Services/FileService.cs
@@ -7,16 +7,22 @@ using System.Threading.Tasks;
 
 namespace Ghini_Bikes.Services
 {
+    /*
+     * LoggerService
+     * 
+     * */
     public class FileService : IFileService
     {
-         private static FileService _instance;
+        private static FileService _instance;
         private static readonly object padlock = new object();
 
         private FileService()
         {
             //System.Console.WriteLine("Constructor called");
         }
-
+        /// <summary>
+        /// 
+        /// </summary>
         public static FileService Instance
         {
             get
@@ -37,8 +43,8 @@ namespace Ghini_Bikes.Services
             private set { }
         }
         private const string directoryPathPhotos = "C:\\Users\\ghine\\Desktop\\Facultate\\Amdaris\\Proiect\\Ghini-Bike\\Ghini-Bikes\\Ghini-Bikes\\Pictures";
-        
-        public  void CopyFile(string filePathSRC)
+
+        public void CopyFile(string filePathSRC)
         {
             var name = Path.GetFileNameWithoutExtension(filePathSRC) + "_copy";
             var extension = Path.GetExtension(filePathSRC);
@@ -49,35 +55,38 @@ namespace Ghini_Bikes.Services
                 File.Copy(filePathSRC, newFilePath);
                 Console.WriteLine("Copied successfully");
             }
-            catch( Exception ex) {  Console.WriteLine(ex.Message); }
-            
+            catch (Exception ex) { Console.WriteLine(ex.Message); }
         }
-        
-        public  void DeleteFile(string filePathSRC)
+
+        public void DeleteFile(string filePathSRC)
         {
+            if (string.IsNullOrEmpty(filePathSRC))
+            {
+                Console.WriteLine("File path is null");
+                return;
+            }
             FileInfo fileInfo = new FileInfo(filePathSRC);
             try
             {
                 fileInfo.Delete();
                 Console.WriteLine("Deleted successfully");
             }
-            catch ( Exception ex ) { Console.WriteLine(ex.Message); }
-          
+            catch (Exception ex) { Console.WriteLine(ex.Message); }
         }
 
-        public  string GetFilePath(string fileName)
+        // private
+        public string GetFilePath(string fileName)
         {
-           var dir = Directory.GetCurrentDirectory();
+            var dir = Directory.GetCurrentDirectory();
             DirectoryInfo directory = new DirectoryInfo(dir);
             DirectoryInfo baseDirectory = directory.Parent.Parent.Parent;
 
-            string pathDirecotry = Path.Combine(baseDirectory.FullName,"Pictures");
-            string filePath = Path.Combine(pathDirecotry,fileName);
-            
+            string pathDirecotry = Path.Combine(baseDirectory.FullName, "Pictures");
+            string filePath = Path.Combine(pathDirecotry, fileName);
+
             //string fullPath = Path.GetFullPath(fileName);
             //Console.WriteLine(filePath);
             return filePath;
-        
         }
 
     }
